Correcion de errores menores producto de merge de ambos branches

+++ backend/controllers/GeneralController.php
@@ -33,24 +33,32 @@ class GeneralController extends Controller
      */
     public function actionIndex()
     {
+        if(Yii::$app->user->can('general-index'))
+        {    
         $searchModel = new GeneralSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
-
         return $this->render('index', [
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
+        }
+        else
+            throw new UnauthorizedHttpException(Yii::t('app', 'You are not authorized to access this view.'));
     }
 
     public function actionSelect()
     {
+        if(Yii::$app->user->can('general-select'))
+        {    
         $searchModel = new GeneralSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
-
         return $this->renderAjax('select', [
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
+        }
+        else
+            throw new UnauthorizedHttpException(Yii::t('app', 'You are not authorized to access this view.'));
     }
 
 
@@ -61,16 +69,26 @@ class GeneralController extends Controller
      */
     public function actionView($id)
     {
+        if(Yii::$app->user->can('general-view'))
+        {    
         return $this->renderAjax('view', [
             'model' => $this->findModel($id),
         ]);
+        }
+        else
+            throw new UnauthorizedHttpException(Yii::t('app', 'You are not authorized to access this view.'));
     }
 
     public function actionAjaxView($id)
     {
+        if(Yii::$app->user->can('general-ajaxview'))
+        {    
         return $this->renderAjax('ajax-view', [
             'model' => $this->findModel($id),
         ]);
+        }
+        else
+            throw new UnauthorizedHttpException(Yii::t('app', 'You are not authorized to access this view.'));
     }
 
     /**
@@ -80,6 +98,8 @@ class GeneralController extends Controller
      */
     public function actionCreate()
     {
+        if(Yii::$app->user->can('general-create'))
+        {
         $model = new General();
         $data=[
             'url'=>null,
@@ -101,10 +121,15 @@ class GeneralController extends Controller
                 'model' => $model,
             ]);
         }
+        }
+        else
+            throw new UnauthorizedHttpException(Yii::t('app', 'You are not authorized to access this view.'));
     }
 
     public function actionEstablish()
     {
+        if(Yii::$app->user->can('general-establish'))
+        {    
         $model = new General();
         $data=[
             'url'=>null,
@@ -126,6 +151,9 @@ class GeneralController extends Controller
                 'model' => $model,
             ]);
         }
+        }
+        else
+            throw new UnauthorizedHttpException(Yii::t('app', 'You are not authorized to access this view.'));
     }
 
 
@@ -137,8 +165,9 @@ class GeneralController extends Controller
      */
     public function actionUpdate($id)
     {
+        if(Yii::$app->user->can('general-update'))
+        {    
         $model = $this->findModel($id);
-
         if ($model->load(Yii::$app->request->post())) {
             $model->save();
             return $this->redirect(['view', 'id' => $model->id]);
@@ -147,6 +176,9 @@ class GeneralController extends Controller
                 'model' => $model,
             ]);
         }
+        }
+        else
+            throw new UnauthorizedHttpException(Yii::t('app', 'You are not authorized to access this view.'));
     }
 
     /**
@@ -157,9 +189,13 @@ class GeneralController extends Controller
      */
     public function actionDelete($id)
     {
+        if(Yii::$app->user->can('general-delete'))
+        {
         $this->findModel($id)->delete();
-
         return $this->redirect(['index']);
+        }
+        else
+            throw new UnauthorizedHttpException(Yii::t('app', 'You are not authorized to access this view.'));
     }
 
     /**
